# docs/openapi.yml
openapi: 3.0.3
info:
  title: Tournament Platform API
  version: v1
  description: |
    API for the Tournament Platform, enabling user management, wallet operations, tournament participation, and administrative functions.
    Key features include secure authentication, real-time match updates, and role-based access control.
  contact:
    name: Support Team
    email: support@example-tournament.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:{port}/api/v1 # Define port in variables or replace directly
    description: Local development server
    variables:
      port:
        default: '3000'
        description: Port for the local server
  - url: https://staging.example-tournament.com/api/v1
    description: Staging server
  - url: https://api.example-tournament.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access Token for authentication. Example: 'Bearer {token}'"
    refreshTokenCookie: # For documenting the refresh token mechanism
      type: apiKey
      in: cookie
      name: jid # Matches JWT_REFRESH_COOKIE_NAME from .env.example
      description: "HttpOnly cookie containing the Refresh Token."
  schemas:
...
    Dispute:
      type: object
      properties:
        id:
          type: string
          format: uuid
        matchId:
          type: string
          format: uuid
        reporterId:
          type: string
          format: uuid
        reason:
          type: string
        status:
          type: string
          enum: [OPEN, UNDER_REVIEW, RESOLVED_PARTICIPANT1_WIN, RESOLVED_PARTICIPANT2_WIN, RESOLVED_REPLAY, CLOSED]
        resolutionDetails:
          type: string
          nullable: true
        moderatorId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginatedDisputes:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Dispute'
    ResolveDisputeRequest:
      type: object
      required:
        - resolutionStatus # E.g., RESOLVED_PARTICIPANT1_WIN
        - resolutionDetails
      properties:
        resolutionStatus:
          type: string
          enum: [RESOLVED_PARTICIPANT1_WIN, RESOLVED_PARTICIPANT2_WIN, RESOLVED_REPLAY, CLOSED_INVALID]
          example: "RESOLVED_PARTICIPANT1_WIN"
        resolutionDetails:
          type: string
          example: "After reviewing evidence, participant 1 is declared the winner."
        # Optional: winningParticipantId if not directly derivable from resolutionStatus
    WithdrawalForAdmin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: number
          format: float
        currency:
          type: string
        status:
          type: string
          enum: [PENDING_APPROVAL, APPROVED, REJECTED, PROCESSING, COMPLETED, FAILED]
        requestedAt:
          type: string
          format: date-time
        withdrawalMethodDetails:
          type: object
        # adminNotes:
        #   type: string
    PaginatedWithdrawalsAdmin:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/WithdrawalForAdmin'
    ApproveWithdrawalRequest:
      type: object
      properties:
        notes: # Optional notes from finance manager
          type: string
          nullable: true
          example: "Approved after verification."
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    DisputeIdPath:
      name: id
      in: path
      required: true
      description: ID of the dispute ticket.
      schema:
        type: string
        format: uuid
    WithdrawalIdPath:
      name: id
      in: path
      required: true
      description: ID of the withdrawal request.
      schema:
        type: string
        format: uuid
...
paths:
...
  /admin/disputes:
    get:
      tags: [Admin - Disputes]
      summary: Get a list of disputes (for DisputeModerator).
      security:
        - bearerAuth: [] # RBAC middleware will check for DisputeModerator role
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [OPEN, UNDER_REVIEW, RESOLVED_PARTICIPANT1_WIN, RESOLVED_PARTICIPANT2_WIN, RESOLVED_REPLAY, CLOSED]
          description: Filter by dispute status.
        - name: tournamentId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter by tournament ID.
      responses:
        '200':
          description: A paginated list of disputes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputes'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden (user does not have DisputeModerator role).
        '500':
          description: Internal server error.
  /admin/disputes/{id}/resolve:
    post:
      tags: [Admin - Disputes]
      summary: Resolve a dispute and determine the winner (for DisputeModerator).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DisputeIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveDisputeRequest'
      responses:
        '200':
          description: Dispute resolved successfully.
          content:
            application/json:
              schema: # Returns the updated dispute ticket
                $ref: '#/components/schemas/Dispute'
        '400':
          description: Validation error.
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '404':
          description: Dispute not found.
        '500':
          description: Internal server error.
  /admin/withdrawals:
    get:
      tags: [Admin - Withdrawals]
      summary: Get a list of withdrawal requests (for FinanceManager).
      security:
        - bearerAuth: [] # RBAC for FinanceManager
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING_APPROVAL, APPROVED, REJECTED, PROCESSING, COMPLETED, FAILED]
          description: Filter by withdrawal status.
        - name: userId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter by user ID.
      responses:
        '200':
          description: A paginated list of withdrawal requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWithdrawalsAdmin'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
  /admin/withdrawals/{id}/approve:
    post:
      tags: [Admin - Withdrawals]
      summary: Approve a withdrawal request (for FinanceManager).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WithdrawalIdPath'
      requestBody:
        required: false # Notes are optional
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveWithdrawalRequest'
      responses:
        '200':
          description: Withdrawal request approved successfully.
          content:
            application/json:
              schema: # Returns the updated withdrawal request
                $ref: '#/components/schemas/WithdrawalForAdmin'
        '400':
          description: Validation error (e.g., request not in approvable state).
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '404':
          description: Withdrawal request not found.
        '500':
          description: Internal server error.
  /admin/withdrawals/{id}/reject: # Added for completeness
    post:
      tags: [Admin - Withdrawals]
      summary: Reject a withdrawal request (for FinanceManager).
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WithdrawalIdPath'
      requestBody:
        required: true # Reason for rejection is usually required
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  example: "Suspected fraudulent activity."
      responses:
        '200':
          description: Withdrawal request rejected successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalForAdmin'
        '400':
          description: Validation error.
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '404':
          description: Withdrawal request not found.
        '500':
          description: Internal server error.

# Placeholder for other routes like users, teams, leaderboards if they have HTTP endpoints
# /users/me (GET) - Get current user's profile
# /teams (POST, GET) - Manage teams
# /leaderboards/{game} (GET) - Get leaderboards
# These are not explicitly detailed in the initial prompt's API list but are in the file structure.
# For now, focusing on the explicitly listed API endpoints.